Completion details, documentation, helper functions
Add code to add additional details to each candidate.  The
location that each candidate came from will be stored as
a text property in the event that company-mode is used.
Also add functions to help identify candidates if a single
type of candidate is not known, ie. the candidate is either
a schema or a table.  Also add some documentation.