Refactor and remove list for query targets
Refactor to use new helper functions for narrowing and to use the
for portion of cl-loop instead of let* within the loops.  The list
that contained all of the targets that were search targets and
resolved names has been removed.  The single hashtable is used now.
Unresolved targets are entered in the hashtable with an empty
string as the value.