Change query to resolve database objects and remove unused functions
Change the query that is used to resolve database objets in an
effort to reduce the number of extra results that are found.  The
query will only match the table name and the schema now.  Remove
two unused functions used to determine if a target string was in
the current buffer and to search backward but not move.